requirements.txt
beautifulsoup4==4.12.2
firebase-admin==6.1.0
requests==2.30.0
pdfplumber==0.9.0
pypdf==3.8.1
tabula-py==2.7.0
